build(folders): :truck: reorganize project structure
reorganize project folders into logical groupings
