+++ bin/cli.js
@@ -21,13 +21,13 @@ import {
     addSecret,
     addUser,
     deleteUser,
-} from '../src/api.js';
+} from '../src/api/api.js';
 
 import initNewUser from '../src/utils/iam/initNewUser.js';
 import deleteIAMUser from '../src/utils/iam/deleteIAMUser.js';
-import init from '../src/utils/aws-init/init.js';
+import init from '../src/aws/init.js';
 
-import injectSecrets from '../src/wrapper.js';
+import injectSecrets from '../src/api/injectSecrets.js';
 
 const cli = new Command();
