+++ src/aws/iam/setup/attachPolicyToUser.js
