+++ src/aws/iam/teardown/detachPoliciesFromUser.js
@@ -1,5 +1,5 @@
 import { IAMClient, DetachUserPolicyCommand } from '@aws-sdk/client-iam';
-import getPolicyARNsFromName from './getPolicyARNsFromName.js';
+import getPolicyARNsFromName from '../../utils/getPolicyARNsFromName.js';
 
 const client = new IAMClient();
