+++ src/aws/utils/doesUserGroupExist.js
