+++ src/aws/utils/generateDBAuthToken.js
