+++ src/aws/utils/getDBArnAndEndpoint.js
