+++ src/aws/utils/getDbResourceId.js
