+++ src/aws/utils/getFormattedResourceArn.js
