+++ src/aws/utils/getUserArn.js
