+++ src/aws/utils/getUsername.js
